"Correct typo in Courses link"
This commit message describes the change in the code diff, which is a correction of a typo in the link for the "Courses" menu item in the Navbar component. The link's href attribute was previously "Courses" with a capital "C", but it has been corrected to "cources" with a lowercase "c".